Add files via upload
Welcome to my GitHub repository! This project focuses on diabetes prediction using machine learning. By analyzing various health indicators, we aim to develop a predictive model that can accurately classify individuals as diabetic or non-diabetic. Through data exploration and model training, we strive to contribute to the advancement of healthcare and early disease detection. Thank you for your interest, and feel free to explore the project!